Extract common (mainly cover related) functionality from BaseMetaTileEntity & BaseMetaPipeEntity
